lb_tokengenerate/requirements.txt
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
